Generalized HDF5 read/write from Parameters class
